requirements.txt
Django==5.1.1
asgiref==3.8.1
channels==4.1.0
djangorestframework==3.15.2
redis==5.0.8
django-grappelli==4.0.1
uvicorn[standard]==0.30.6
daphne==2.5.0
channels_redis==4.2.0
dj_database_url==2.2.0
psycopg2-binary==2.9.9
django_redis==5.4.0